+++ moor_generator/lib/src/writer/tables/table_writer.dart
@@ -5,174 +5,15 @@ import 'package:moor_generator/src/utils/string_escaper.dart';
 import 'package:moor_generator/writer.dart';
 import 'package:sqlparser/sqlparser.dart';
 
-class TableWriter {
-  final MoorTable table;
-  final Scope scope;
+/// Common writer for tables or views.
+///
+/// Both classes need to generate column getters and a mapping function.
+abstract class TableOrViewWriter {
+  MoorEntityWithResultSet get tableOrView;
+  StringBuffer get buffer;
 
-  StringBuffer _buffer;
-
-  TableWriter(this.table, this.scope);
-
-  bool get _skipVerification =>
-      scope.writer.options.skipVerificationCode ||
-      scope.generationOptions.isGeneratingForSchema;
-
-  void writeInto() {
-    writeDataClass();
-    writeTableInfoClass();
-  }
-
-  void writeDataClass() {
-    if (!table.hasExistingRowClass &&
-        scope.generationOptions.writeDataClasses) {
-      DataClassWriter(table, scope.child()).write();
-    }
-
-    if (scope.generationOptions.writeCompanions) {
-      UpdateCompanionWriter(table, scope.child()).write();
-    }
-  }
-
-  void writeTableInfoClass() {
-    _buffer = scope.leaf();
-
-    if (scope.generationOptions.isGeneratingForSchema) {
-      // Write a small table header without data class
-      _buffer.write('class ${table.tableInfoName} extends Table with '
-          'TableInfo');
-      if (table.isVirtualTable) {
-        _buffer.write(', VirtualTableInfo');
-      }
-    } else {
-      // Regular generation, write full table class
-      final dataClass = table.dartTypeName;
-      final tableDslName = table.fromClass?.name ?? 'Table';
-
-      // class UsersTable extends Users implements TableInfo<Users, User> {
-      final typeArgs = '<${table.tableInfoName}, $dataClass>';
-      _buffer.write('class ${table.tableInfoName} extends $tableDslName with '
-          'TableInfo$typeArgs ');
-
-      if (table.isVirtualTable) {
-        _buffer.write(', VirtualTableInfo$typeArgs ');
-      }
-    }
-
-    _buffer
-      ..write('{\n')
-      // write a GeneratedDatabase reference that is set in the constructor
-      ..write('final GeneratedDatabase _db;\n')
-      ..write('final ${scope.nullableType('String')} _alias;\n')
-      ..write('${table.tableInfoName}(this._db, [this._alias]);\n');
-
-    // Generate the columns
-    for (final column in table.columns) {
-      _writeColumnVerificationMeta(column);
-      _writeColumnGetter(column);
-    }
-
-    // Generate $columns, $tableName, asDslTable getters
-    final columnsWithGetters =
-        table.columns.map((c) => c.dartGetterName).join(', ');
-
-    _buffer
-      ..write('@override\nList<GeneratedColumn> get \$columns => '
-          '[$columnsWithGetters];\n')
-      ..write('@override\n${table.tableInfoName} get asDslTable => this;\n')
-      ..write('@override\nString get \$tableName => '
-          '_alias ?? \'${table.sqlName}\';\n')
-      ..write(
-          '@override\nfinal String actualTableName = \'${table.sqlName}\';\n');
-
-    _writeValidityCheckMethod();
-    _writePrimaryKeyOverride();
-
-    _writeMappingMethod();
-    // _writeReverseMappingMethod();
-
-    _writeAliasGenerator();
-
-    _writeConvertersAsStaticFields();
-    _overrideFieldsIfNeeded();
-
-    // close class
-    _buffer.write('}');
-  }
-
-  void _writeConvertersAsStaticFields() {
-    for (final converter in table.converters) {
-      final typeName = converter.converterNameInCode(scope.generationOptions);
-      final code = converter.expression;
-      _buffer.write('static $typeName ${converter.fieldName} = $code;');
-    }
-  }
-
-  void _writeMappingMethod() {
-    if (!scope.generationOptions.writeDataClasses) {
-      final nullableString = scope.nullableType('String');
-      _buffer.writeln('''
-        @override
-        Never map(Map<String, dynamic> data, {$nullableString tablePrefix}) {
-          throw UnsupportedError('TableInfo.map in schema verification code');
-        }
-      ''');
-      return;
-    }
-
-    final dataClassName = table.dartTypeName;
-
-    _buffer.write('@override\n$dataClassName map(Map<String, dynamic> data, '
-        '{${scope.nullableType('String')} tablePrefix}) {\n');
-
-    if (table.hasExistingRowClass) {
-      _buffer.write('final effectivePrefix = '
-          "tablePrefix != null ? '\$tablePrefix.' : '';");
-
-      final info = table.existingRowClass;
-      final positionalToIndex = <MoorColumn, int>{};
-      final named = <MoorColumn, String>{};
-
-      final parameters = info.constructor.parameters;
-      info.mapping.forEach((column, parameter) {
-        if (parameter.isNamed) {
-          named[column] = parameter.name;
-        } else {
-          positionalToIndex[column] = parameters.indexOf(parameter);
-        }
-      });
-
-      // Sort positional columns by the position of their respective parameter
-      // in the constructor.
-      final positional = positionalToIndex.keys.toList()
-        ..sort((a, b) => positionalToIndex[a].compareTo(positionalToIndex[b]));
-
-      final writer = RowMappingWriter(
-        positional,
-        named,
-        table,
-        scope.generationOptions,
-        dbName: '_db',
-      );
-
-      final classElement = info.targetClass;
-      final ctor = info.constructor;
-      _buffer..write('return ')..write(classElement.name);
-      if (ctor.name != null && ctor.name.isNotEmpty) {
-        _buffer..write('.')..write(ctor.name);
-      }
-
-      writer.writeArguments(_buffer);
-      _buffer.write(';\n');
-    } else {
-      // Use default .fromData constructor in the moor-generated data class
-      _buffer.write('return $dataClassName.fromData(data, _db, '
-          "prefix: tablePrefix != null ? '\$tablePrefix.' : null);\n");
-    }
-
-    _buffer.write('}\n');
-  }
-
-  void _writeColumnGetter(MoorColumn column) {
+  void writeColumnGetter(
+      MoorColumn column, GenerationOptions options, bool isOverride) {
     final isNullable = column.nullable;
     final additionalParams = <String, String>{};
     final expressionBuffer = StringBuffer();
@@ -206,7 +47,7 @@ class TableWriter {
 
     expressionBuffer
       ..write('return ${column.implColumnTypeName}')
-      ..write("('${column.name.name}', \$tableName, $isNullable, ");
+      ..write("('${column.name.name}', aliasedName, $isNullable, ");
 
     var first = true;
     additionalParams.forEach((name, value) {
@@ -228,20 +69,199 @@ class TableWriter {
     expressionBuffer.write(';');
 
     writeMemoizedGetterWithBody(
-      buffer: _buffer,
+      buffer: buffer,
       getterName: column.dartGetterName,
       returnType: column.implColumnTypeName,
       code: expressionBuffer.toString(),
-      // don't override on custom tables because we only override the column
-      // when the base class is user defined
-      hasOverride: !table.isFromSql,
-      options: scope.generationOptions,
+      hasOverride: isOverride,
+      options: options,
     );
+  }
+
+  void writeGetColumnsOverride() {
+    final columnsWithGetters =
+        tableOrView.columns.map((c) => c.dartGetterName).join(', ');
+    buffer.write('@override\nList<GeneratedColumn> get \$columns => '
+        '[$columnsWithGetters];\n');
+  }
+
+  void writeAsDslTable() {
+    buffer.write(
+        '@override\n${tableOrView.entityInfoName} get asDslTable => this;\n');
+  }
+}
+
+class TableWriter extends TableOrViewWriter {
+  final MoorTable table;
+  final Scope scope;
+
+  @override
+  StringBuffer buffer;
+
+  @override
+  MoorTable get tableOrView => table;
+
+  TableWriter(this.table, this.scope);
+
+  bool get _skipVerification =>
+      scope.writer.options.skipVerificationCode ||
+      scope.generationOptions.isGeneratingForSchema;
+
+  void writeInto() {
+    writeDataClass();
+    writeTableInfoClass();
+  }
+
+  void writeDataClass() {
+    if (!table.hasExistingRowClass &&
+        scope.generationOptions.writeDataClasses) {
+      DataClassWriter(table, scope.child()).write();
+    }
+
+    if (scope.generationOptions.writeCompanions) {
+      UpdateCompanionWriter(table, scope.child()).write();
+    }
+  }
+
+  void writeTableInfoClass() {
+    buffer = scope.leaf();
+
+    if (!scope.generationOptions.writeDataClasses) {
+      // Write a small table header without data class
+      buffer.write('class ${table.entityInfoName} extends Table with '
+          'TableInfo');
+      if (table.isVirtualTable) {
+        buffer.write(', VirtualTableInfo');
+      }
+    } else {
+      // Regular generation, write full table class
+      final dataClass = table.dartTypeName;
+      final tableDslName = table.fromClass?.name ?? 'Table';
+
+      // class UsersTable extends Users implements TableInfo<Users, User> {
+      final typeArgs = '<${table.entityInfoName}, $dataClass>';
+      buffer.write('class ${table.entityInfoName} extends $tableDslName with '
+          'TableInfo$typeArgs ');
+
+      if (table.isVirtualTable) {
+        buffer.write(', VirtualTableInfo$typeArgs ');
+      }
+    }
+
+    buffer
+      ..write('{\n')
+      // write a GeneratedDatabase reference that is set in the constructor
+      ..write('final GeneratedDatabase _db;\n')
+      ..write('final ${scope.nullableType('String')} _alias;\n')
+      ..write('${table.entityInfoName}(this._db, [this._alias]);\n');
+
+    // Generate the columns
+    for (final column in table.columns) {
+      _writeColumnVerificationMeta(column);
+      // Only add an @override to a column getter if we're overriding the column
+      // from a Dart DSL class.
+      writeColumnGetter(column, scope.generationOptions, !table.isFromSql);
+    }
+
+    // Generate $columns, $tableName, asDslTable getters
+    writeGetColumnsOverride();
+    buffer
+      ..write('@override\nString get aliasedName => '
+          '_alias ?? \'${table.sqlName}\';\n')
+      ..write(
+          '@override\n String get actualTableName => \'${table.sqlName}\';\n');
+
+    _writeValidityCheckMethod();
+    _writePrimaryKeyOverride();
+
+    _writeMappingMethod();
+    // _writeReverseMappingMethod();
+
+    _writeAliasGenerator();
+
+    _writeConvertersAsStaticFields();
+    _overrideFieldsIfNeeded();
+
+    // close class
+    buffer.write('}');
+  }
+
+  void _writeConvertersAsStaticFields() {
+    for (final converter in table.converters) {
+      final typeName = converter.converterNameInCode(scope.generationOptions);
+      final code = converter.expression;
+      buffer.write('static $typeName ${converter.fieldName} = $code;');
+    }
+  }
+
+  void _writeMappingMethod() {
+    if (!scope.generationOptions.writeDataClasses) {
+      final nullableString = scope.nullableType('String');
+      buffer.writeln('''
+        @override
+        Never map(Map<String, dynamic> data, {$nullableString tablePrefix}) {
+          throw UnsupportedError('TableInfo.map in schema verification code');
+        }
+      ''');
+      return;
+    }
+
+    final dataClassName = table.dartTypeName;
+
+    buffer.write('@override\n$dataClassName map(Map<String, dynamic> data, '
+        '{${scope.nullableType('String')} tablePrefix}) {\n');
+
+    if (table.hasExistingRowClass) {
+      buffer.write('final effectivePrefix = '
+          "tablePrefix != null ? '\$tablePrefix.' : '';");
+
+      final info = table.existingRowClass;
+      final positionalToIndex = <MoorColumn, int>{};
+      final named = <MoorColumn, String>{};
+
+      final parameters = info.constructor.parameters;
+      info.mapping.forEach((column, parameter) {
+        if (parameter.isNamed) {
+          named[column] = parameter.name;
+        } else {
+          positionalToIndex[column] = parameters.indexOf(parameter);
+        }
+      });
+
+      // Sort positional columns by the position of their respective parameter
+      // in the constructor.
+      final positional = positionalToIndex.keys.toList()
+        ..sort((a, b) => positionalToIndex[a].compareTo(positionalToIndex[b]));
+
+      final writer = RowMappingWriter(
+        positional,
+        named,
+        table,
+        scope.generationOptions,
+        dbName: '_db',
+      );
+
+      final classElement = info.targetClass;
+      final ctor = info.constructor;
+      buffer..write('return ')..write(classElement.name);
+      if (ctor.name != null && ctor.name.isNotEmpty) {
+        buffer..write('.')..write(ctor.name);
+      }
+
+      writer.writeArguments(buffer);
+      buffer.write(';\n');
+    } else {
+      // Use default .fromData constructor in the moor-generated data class
+      buffer.write('return $dataClassName.fromData(data, _db, '
+          "prefix: tablePrefix != null ? '\$tablePrefix.' : null);\n");
+    }
+
+    buffer.write('}\n');
   }
 
   void _writeColumnVerificationMeta(MoorColumn column) {
     if (!_skipVerification) {
-      _buffer
+      buffer
         ..write('final VerificationMeta ${_fieldNameForColumnMeta(column)} = ')
         ..write("const VerificationMeta('${column.dartGetterName}');\n");
     }
@@ -250,7 +270,7 @@ class TableWriter {
   void _writeValidityCheckMethod() {
     if (_skipVerification) return;
 
-    _buffer
+    buffer
       ..write('@override\nVerificationContext validateIntegrity'
           '(Insertable<${table.dartTypeName}> instance, '
           '{bool isInserting = false}) {\n')
@@ -268,13 +288,13 @@ class TableWriter {
       if (column.typeConverter != null) {
         // dont't verify custom columns, we assume that the user knows what
         // they're doing
-        _buffer.write(
+        buffer.write(
             'context.handle($metaName, const VerificationResult.success());');
         continue;
       }
 
       final columnNameString = asDartLiteral(column.name.name);
-      _buffer
+      buffer
         ..write('if (data.containsKey($columnNameString)) {\n')
         ..write('context.handle('
             '$metaName, '
@@ -283,13 +303,13 @@ class TableWriter {
         ..write('}');
 
       if (table.isColumnRequiredForInsert(column)) {
-        _buffer
+        buffer
           ..write(' else if (isInserting) {\n')
           ..write('context.missing($metaName);\n')
           ..write('}\n');
       }
     }
-    _buffer.write('return context;\n}\n');
+    buffer.write('return context;\n}\n');
   }
 
   String _fieldNameForColumnMeta(MoorColumn column) {
@@ -297,7 +317,7 @@ class TableWriter {
   }
 
   void _writePrimaryKeyOverride() {
-    _buffer.write('@override\nSet<GeneratedColumn> get \$primaryKey => ');
+    buffer.write('@override\nSet<GeneratedColumn> get \$primaryKey => ');
     var primaryKey = table.fullPrimaryKey;
 
     // If there is an auto increment column, that forms the primary key. The
@@ -306,27 +326,27 @@ class TableWriter {
     primaryKey ??= table.columns.where((c) => c.hasAI).toSet();
 
     if (primaryKey.isEmpty) {
-      _buffer.write('<GeneratedColumn>{};');
+      buffer.write('<GeneratedColumn>{};');
       return;
     }
 
-    _buffer.write('{');
+    buffer.write('{');
     final pkList = primaryKey.toList();
     for (var i = 0; i < pkList.length; i++) {
       final pk = pkList[i];
 
-      _buffer.write(pk.dartGetterName);
+      buffer.write(pk.dartGetterName);
       if (i != pkList.length - 1) {
-        _buffer.write(', ');
+        buffer.write(', ');
       }
     }
-    _buffer.write('};\n');
+    buffer.write('};\n');
   }
 
   void _writeAliasGenerator() {
-    final typeName = table.tableInfoName;
+    final typeName = table.entityInfoName;
 
-    _buffer
+    buffer
       ..write('@override\n')
       ..write('$typeName createAlias(String alias) {\n')
       ..write('return $typeName(_db, alias);')
@@ -336,23 +356,21 @@ class TableWriter {
   void _overrideFieldsIfNeeded() {
     if (table.overrideWithoutRowId != null) {
       final value = table.overrideWithoutRowId ? 'true' : 'false';
-      _buffer
-        ..write('@override\n')
-        ..write('bool get withoutRowId => $value;\n');
+      buffer..write('@override\n')..write('bool get withoutRowId => $value;\n');
     }
 
     if (table.overrideTableConstraints != null) {
       final value =
           table.overrideTableConstraints.map(asDartLiteral).join(', ');
 
-      _buffer
+      buffer
         ..write('@override\n')
         ..write('List<String> get customConstraints => const [$value];\n');
     }
 
     if (table.overrideDontWriteConstraints != null) {
       final value = table.overrideDontWriteConstraints ? 'true' : 'false';
-      _buffer
+      buffer
         ..write('@override\n')
         ..write('bool get dontWriteConstraints => $value;\n');
     }
@@ -362,7 +380,7 @@ class TableWriter {
       final stmt = declaration.creatingStatement as CreateVirtualTableStatement;
       final moduleAndArgs = asDartLiteral(
           '${stmt.moduleName}(${stmt.argumentContent.join(', ')})');
-      _buffer
+      buffer
         ..write('@override\n')
         ..write('String get moduleAndArgs => $moduleAndArgs;\n');
     }